refactor: move shared Contact* components out of pages
