+++ src/components/contact/ContactPaymentsHistory.vue
@@ -36,14 +36,14 @@ import { Paginated, PaymentStatus } from '@beabee/beabee-common';
 import { useI18n } from 'vue-i18n';
 import { computed, ref, watchEffect } from 'vue';
 
-import AppTable from '../../../table/AppTable.vue';
-import AppPagination from '../../../AppPagination.vue';
-import { formatLocale } from '../../../../utils/dates/locale-date-formats';
+import AppTable from '../table/AppTable.vue';
+import AppPagination from '../AppPagination.vue';
+import { formatLocale } from '../../utils/dates/locale-date-formats';
 
-import { fetchPayments } from '../../../../utils/api/member';
-import { GetPaymentData } from '../../../../utils/api/api.interface';
-import { Header, SortType } from '../../../table/table.interface';
-import AppHeading from '../../../AppHeading.vue';
+import { fetchPayments } from '../../utils/api/member';
+import { GetPaymentData } from '../../utils/api/api.interface';
+import { Header, SortType } from '../table/table.interface';
+import AppHeading from '../AppHeading.vue';
 
 const { t, n } = useI18n();
