+++ src/components/contact/ContactUpdateBasicInformation.vue
@@ -31,7 +31,7 @@
 <script lang="ts" setup>
 import { computed } from 'vue';
 import { useI18n } from 'vue-i18n';
-import AppInput from './forms/AppInput.vue';
+import AppInput from '../forms/AppInput.vue';
 
 const { t } = useI18n();
