+++ src/pages/profile/contribution/index.vue
@@ -35,7 +35,7 @@ meta:
       <CancelContribution :contribution="contribution" />
     </div>
     <div>
-      <PaymentsHistory id="me" class="lg:ml-10" />
+      <ContactPaymentsHistory id="me" class="lg:ml-10" />
     </div>
   </div>
 </template>
@@ -55,7 +55,7 @@ import CancelContribution from '../../../components/pages/profile/contribution/C
 import PaymentSource from '../../../components/pages/profile/contribution/PaymentSource.vue';
 import PageTitle from '../../../components/PageTitle.vue';
 import AppAlert from '../../../components/AppAlert.vue';
-import PaymentsHistory from '../../../components/pages/profile/contribution/PaymentsHistory.vue';
+import ContactPaymentsHistory from '../../../components/contact/ContactPaymentsHistory.vue';
 import { currentUser } from '../../../store';
 import UpdateContribution from '../../../components/pages/profile/contribution/UpdateContribution.vue';
 import { ContributionInfo } from '../../../utils/api/api.interface';
